Fixed LaserStop Bounding
Fixed #8. Turns out rotating by 180 degrees causes issues with the
rotation calculations. Setting it to 180.1 seems to fix everything. Also
added saving of data upon return to main menu from options menu.

+++ src/edu/ncsu/feddgame/render/Model.java
@@ -198,8 +198,8 @@ public class Model {
 	 */
 	public void rotate(float angle){
 			for (int i = 0; i < sideCount; i++){
-				float newX = (float) (xOffset + (this.vertices[i * 3]-xOffset)*Math.cos(angle) - (this.vertices[i * 3 + 1]-yOffset)*Math.sin(angle));
-				float newY = (float) (yOffset + (this.vertices[i * 3]-xOffset)*Math.sin(angle) + (this.vertices[i * 3 + 1]-yOffset)*Math.cos(angle));
+				float newX = (float) (xOffset + (this.vertices[i * 3]-xOffset)*(float)Math.cos(angle) - (this.vertices[i * 3 + 1]-yOffset)*(float)Math.sin(angle));
+				float newY = (float) (yOffset + (this.vertices[i * 3]-xOffset)*(float)Math.sin(angle) + (this.vertices[i * 3 + 1]-yOffset)*(float)Math.cos(angle));
 				vertices[i * 3] = newX;
 				vertices[i * 3 + 1] = newY;
 			}
